Fix : 오타수정 entitiy > entity (#50)
