+++ order-server/src/main/java/com/inbobwetrust/model/entity/Delivery.java
@@ -1,4 +1,4 @@
-package com.inbobwetrust.model.entitiy;
+package com.inbobwetrust.model.entity;
 
 import lombok.Builder;
 import lombok.Getter;
